# swagger.yaml
openapi: 3.0.0
info:
  description: 'DaSCR-board API specification'
  version: 'v1'
  title: 'DaSCR-Board API'
servers:
  - url: 'http://localhost:8000/api'
    variables: {}
    description: Default Config or Docker
...
paths:
  '/debug/{id}/redirect':
    get:
      parameters:
        - $ref: '#/components/parameters/GameID'
      tags:
        - Debug
      summary: Trigger websocket redirect
      responses:
        '200':
          description: successful operation
  '/debug/{id}/update':
    get:
      parameters:
        - $ref: '#/components/parameters/GameID'
      tags:
        - Debug
      summary: Trigger websocket update
      responses:
        '200':
          description: successful operation
  '/debug/{id}/display':
    get:
      tags:
        - Debug
      summary: Get a specific game by id (not stripped with full data)
      parameters:
        - $ref: '#/components/parameters/GameID'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Base'
        "404":
          description: game with id not found
...
components:
  schemas:
...
    Player:
      type: object
      properties:
        UID:
          type: integer
          example: 1
        Name:
          type: string
          example: "Patrick"
        Nickname:
          type: string
          example: "The l33t c0der"
        Image:
          type: string
          example: "images/static_7.png"
        ThrowRounds:
          type: array
          items:
            $ref: "#/components/schemas/ThrowRound"
        TotalThrowCount:
          type: integer
          example: 6
        Score:
          type: object
          $ref: "#/components/schemas/Score"
        LastThrows:
          type: array
          items:
            $ref: "#/components/schemas/Throw"
        ThrowSum:
          type: integer
          example: 81
        Average:
          type: number
          example: 45.89
...
    Base:
      type: object
      properties:
        Base:
          type: object
          properties:
            UID:
              type: string
              example: "dascr"
            Game:
              type: string
              example: "x01"
            Player:
              type: array
              items:
                $ref: "#/components/schemas/Player"
            Variant:
              type: string
              example: "501"
            In:
              type: string
              example: "straight"
            Out:
              type: string
              example: "double"
            Elimination:
              type: boolean
              example: false
            ActivePlayer:
              type: integer
              example: 1
            ThrowRound:
              type: integer
              example: 2
            GameState:
              type: string
              example: "THROW"
            Message:
              type: string
              example: "-"
            Settings:
              type: object
              properties:
                Sound:
                  type: boolean
                  example: true
                Podium:
                  type: boolean
                  example: false
                Autoswitch:
                  type: boolean
                  example: false
                Punisher:
                  type: boolean
                  example: false
            UndoLog:
              type: array
              $ref: "#/components/schemas/Sequence"
            Podium:
              type: array
              items:
                $ref: "#/components/schemas/Player"
            CricketController:
              type: object
              properties:
                NumberClosed:
                  type: array
                  example: [false, false, false, false, false, false, false]
                Numbers:
                  type: array
                  example: [15,16,17,18,19,20,25]
                NumberRevealed:
                  type: array
                  example: [false, false, false, false, false, false, false]
                Random:
                  type: boolean
                  example: false
                Ghost:
                  type: boolean
                  example: false
    ThrowRound:
      type: object
      properties:
        Round:
          type: integer
          example: 1
        Done:
          type: boolean
          example: false
        Throws:
          type: array
          items:
            $ref: "#/components/schemas/Throw"
    Score:
      type: object
      properties:
        Score:
          type: integer
          example: 321
        ParkScore:
          type: integer
          example: 381
        InitialScore:
          type: integer
          example: 501
        Numbers:
          type: array
          example: [0, 0, 0, 0, 0, 0, 0]
        Closed:
          type: array
          example: [false, false, false, false, false, false, false]
        CurrentNumber:
          type: integer
          example: 15
        Split:
          type: boolean
          example: false
        Hit:
          type: boolean
          example: false
    Throw:
      type: object
      properties:
        Number:
          type: integer
          example: 20
        Modifier:
          type: integer
          example: 3
    Sequence:
      type: object
      properties:
        Sequence:
          type: integer
          example: 1
        Action:
          type: array
          items:
            $ref: "#/components/schemas/Action"
    Action:
      type: object
      properties:
        Number:
          type: integer
          example: 1
        Action:
          type: string
          example: "CREATEGAME"
        Points:
          type: integer
          example: 0
        RoundNumber:
          type: integer
          example: 0
        Player:
          type: object
          $ref: "#/components/schemas/Player"
        GameID:
          type: string
          example: ""
        NumberIndex:
          type: integer
          example: 0
        Modifier:
          type: integer
          example: 0
        PreviousGameState:
          type: string
          example: ""
        PreviousScore:
          type: integer
          example: 0
        PreviousParkScore:
          type: integer
          example: 0
        PreviousPlayerIndex:
          type: integer
          example: 0
        PreviousThrowSum:
          type: integer
          example: 0
        PreviousAverage:
          type: integer
          example: 0
        PreviousLastThree:
          type: array
          $ref: "#/components/schemas/Throw"
        PreviousMessage:
          type: string
          example: ""
        PreviousNumberToHit:
          type: integer
          example: 0
...
  parameters:
    GameID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 'Unique Game ID'
...